Mu.GRpc.Client.Optics

Description

For further information over initialization of the connection, consult the http2-client-grpc docs.

Initialization of the gRPC client

data GRpcConnection (s :: Package') (p :: GRpcMessageProtocol)

Represents a connection to the service s.

initGRpc

Arguments

:: GrpcClientConfig
gRPC configuration

-> Proxy p

-> IO (Either ClientError (GRpcConnection s p))

Initializes a connection to a gRPC server. Usually the service you are connecting to is inferred from the usage later on. However, it can also be made explicit by using

initGRpc config @Service

data GRpcMessageProtocol

Constructors

MsgProtoBuf

MsgAvro

msgProtoBuf :: Proxy MsgProtoBuf

msgAvro :: Proxy MsgAvro

data GrpcClientConfig

grpcClientConfigSimple :: HostName -> PortNumber -> UseTlsOrNot -> GrpcClientConfig

Request arguments and responses

data CompressMode

data GRpcReply a

Constructors

GRpcTooMuchConcurrency TooMuchConcurrency

GRpcErrorCode ErrorCode

GRpcErrorString String

GRpcClientError ClientError

GRpcOk a

Re-exported for convenience