Mu.GRpc.Client.Record

Description

For further information over initialization of the connection, consult the http2-client-grpc docs.

Initialization of the gRPC client

data GrpcClient

data GrpcClientConfig

grpcClientConfigSimple :: HostName -> PortNumber -> UseTlsOrNot -> GrpcClientConfig

setupGrpcClient' :: GrpcClientConfig -> IO (Either ClientError GrpcClient)

Fill and generate the Haskell record of functions

buildService :: forall (pro :: GRpcMessageProtocol) (pkg :: Package') (s :: Symbol) (p :: Symbol) t (pkgName :: Symbol) (ss :: [Service']) (anns :: [ServiceAnnotation]) (ms :: [Method Symbol Symbol]). (pkg ~ Package (Just pkgName) ss, LookupService ss s ~ Service s anns ms, Generic t, BuildService pro pkgName s p ms (Rep t)) => GrpcClient -> t

Fills in a Haskell record of functions with the corresponding calls to gRPC services from a Mu Service declaration.

data GRpcMessageProtocol

Constructors

MsgProtoBuf

MsgAvro

data CompressMode

Constructors

Compressed

Uncompressed

data GRpcReply a

Constructors

GRpcTooMuchConcurrency TooMuchConcurrency

GRpcErrorCode ErrorCode

GRpcErrorString String

GRpcClientError ClientError

GRpcOk a

generateRecordFromService :: String -> String -> Namer -> Name -> Q [Dec]

Generate the plain Haskell record corresponding to a Mu Service definition, and a concrete implementation of buildService for that record.